--- src/BooksStore.Debug.Console/Program.cs
@@ -1,5 +1,0 @@
-﻿using BooksStore.Core.BookAggregate;
-
-var str = "debug....";
-Console.WriteLine(str);
-Console.WriteLine(new Book("a new book", 3.33m));
