--- src/BooksStore.SharedKernel/BaseDomainEvent.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace SharedKernel;
-
-public abstract class BaseDomainEvent : INotification
-{
-    public DateTime DateOccurred { get; set; } = DateTime.UtcNow;
-}
