--- src/BooksStore.SharedKernel/Interfaces/IAggregateRoot.cs
@@ -1,5 +1,0 @@
-﻿namespace SharedKernel.Interfaces;
-
-public interface IAggregateRoot
-{
-}
