--- src/BooksStore.SharedKernel/Interfaces/IReadRepository.cs
@@ -1,7 +1,0 @@
-﻿using Ardalis.Specification;
-
-namespace SharedKernel.Interfaces;
-
-public interface IReadRepository<T> : IReadRepositoryBase<T> where T : BaseEntity, IAggregateRoot
-{
-}
