--- src/BooksStore.SharedKernel/Interfaces/IRepository.cs
@@ -1,7 +1,0 @@
-﻿using Ardalis.Specification;
-
-namespace SharedKernel.Interfaces;
-
-public interface IRepository<T> : IRepositoryBase<T> where T : BaseEntity, IAggregateRoot
-{
-}
