+++ src/Contracts/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace Contracts;
+public class Class1
+{
+
+}
