+++ src/Core/BooksStore.Application/Abstractions/Messaging/ICommand.cs
