+++ src/Core/BooksStore.Application/Abstractions/Messaging/IQuery.cs
