+++ src/Core/BooksStore.Application/Abstractions/Messaging/IQueryHandler.cs
