+++ src/Core/BooksStore.Application/Books/Commands/CreateBookCommand.cs
