+++ src/Core/BooksStore.Application/Books/Commands/CreateBookCommandHandler.cs
