+++ src/Core/BooksStore.Application/Books/Queries/BookResponse.cs
