+++ src/Core/BooksStore.Application/Books/Queries/GetBookByIdQuery.cs
