+++ src/Core/BooksStore.Application/Books/Queries/GetBookQueryHandler.cs
