+++ src/Core/BooksStore.Domain/Abstractions/IbookRepository.cs
