+++ src/Core/BooksStore.Domain/Class1.cs
