+++ src/Core/BooksStore.Domain/Exceptions/BookNotFoundException.cs
