+++ src/Core/SharedKernel/BaseDomainEvent.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+
+namespace SharedKernel;
+
+public abstract class BaseDomainEvent : INotification
+{
+    public DateTime DateOccurred { get; set; } = DateTime.UtcNow;
+}
