+++ src/Core/SharedKernel/Exceptions/NotFoundException.cs
