+++ src/Core/SharedKernel/Interfaces/IAggregateRoot.cs
@@ -1,0 +1,5 @@
+﻿namespace SharedKernel.Interfaces;
+
+public interface IAggregateRoot
+{
+}
