+++ src/Core/SharedKernel/Interfaces/IReadRepository.cs
@@ -1,0 +1,7 @@
+﻿using Ardalis.Specification;
+
+namespace SharedKernel.Interfaces;
+
+public interface IReadRepository<T> : IReadRepositoryBase<T> where T : BaseEntity, IAggregateRoot
+{
+}
