+++ src/Core/SharedKernel/Interfaces/IRepository.cs
@@ -1,0 +1,7 @@
+﻿using Ardalis.Specification;
+
+namespace SharedKernel.Interfaces;
+
+public interface IRepository<T> : IRepositoryBase<T> where T : BaseEntity, IAggregateRoot
+{
+}
