+++ src/Core/SharedKernel/SharedKernel.csproj
