+++ src/External/BooksStore.Debug.Console/BooksStore.Debug.Console.csproj
@@ -1,17 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BooksStore.Application\BooksStore.Application.csproj" />
+    <ProjectReference Include="..\..\Core\BooksStore.Application\BooksStore.Application.csproj" />
   </ItemGroup>
 
 </Project>
