+++ src/External/BooksStore.Debug.Console/Program.cs
@@ -1,0 +1,6 @@
+﻿
+using BooksStore.Domain.Entities;
+
+var str = "debug....";
+Console.WriteLine(str);
+Console.WriteLine(new Book("a new book")) ;
