+++ src/External/BooksStore.Infrastructure/ApplicationDbContext.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 
 namespace BooksStore.Infrastructure;
+
 public sealed class ApplicationDbContext : DbContext, IUnitOfWork
 {
     public ApplicationDbContext(DbContextOptions options) : base(options) { }
