+++ src/External/BooksStore.Infrastructure/Configurations/BookConfiguration.cs
