+++ src/External/BooksStore.Infrastructure/Repositories/BookRepository.cs
