+++ tests/Architecture.Tests/ArchitectureTests.cs
